+++ src/config/structs.rs
@@ -173,6 +173,8 @@ const fn is_zero(n: &usize) -> bool {
 pub struct Profile {
     #[serde(flatten)]
     pub filters: Filters,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub imports: Vec<PathBuf>,
     #[serde(default, skip_serializing_if = "HashMap::is_empty")]
     pub mods: HashMap<String, Source>,
     #[serde(default, skip_serializing_if = "HashMap::is_empty")]
@@ -197,6 +199,7 @@ impl Profile {
                 },
                 ..Filters::empty()
             },
+            imports: Vec::new(),
             mods: HashMap::new(),
             shaders: HashMap::new(),
             modpacks: HashMap::new(),