+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Restructuring, refactors and breaking changes
 
 - Added embedded profiles
+- Added `imports` to `Profile` struct
 - Only store the `.minecraft` dir and find the other folder locations at runtime
 - Made `add` display the project type
 - Some miscellaneous refactors